Create new file for rcade to check for release versions
This file replaces the old "versions" file for newer releases. Upgrade processes and filenames have changed for some board models, so newer releases will utilize this file when checking for available updates.

The old file will no longer be updated and will continue to allow older versions to upgrade to the version that switches the check to this file.